+++ CHANGELOG.md
@@ -96,6 +96,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Update Plex movie/other video plot ("summary") during library deep scan
 - Fix compatibility with ffmpeg 7.2+ when using NVIDIA accel and 10-bit source content
 - Fix some NVIDIA edge cases when media servers don't provide video bit depth information
+- Fix VAAPI tonemap failure
+- Fix green bars after VAAPI tonemap
 
 ## [25.2.0] - 2025-06-24
 ### Added